Everything.
- Create Markdown component
- Make .grid children at least 250px wide
- Move markdown and prism css files to lib
- Add global label and svg.bi style
- Move all the pages to the top folder again
- Create new `/content` route with blog, projects and photos as children
- Keep `/projects/tanuki` as backup
