+++ src/lib/style/main.css
@@ -70,7 +70,7 @@ body {
 
 .grid {
 	display: grid;
-	grid-template-columns: repeat(auto-fit, minmax(200px, 1fr));
+	grid-template-columns: repeat(auto-fit, minmax(250px, 1fr));
 	gap: .5em;
 }
 
@@ -112,10 +112,6 @@ button,
 textarea,
 select {
 	margin: 2px;
-}
-
-label {
-	cursor: pointer;
 }
