+++ src/lib/style/markdown.css
@@ -4,7 +4,7 @@
 		--color-fg-default: #c9d1d9;
 		--color-fg-muted: #8b949e;
 		--color-fg-subtle: #484f58;
-		--color-canvas-default: #0d1117;
+		/* --color-canvas-default: #0d1117; */
 		--color-canvas-subtle: #161b22;
 		--color-border-default: #30363d;
 		--color-border-muted: #21262d;
