+++ src/lib/style/misc.css
@@ -1,4 +1,15 @@
 /* Some other tags */
+label {
+	display: block;
+	margin-block-start: 2em;
+	margin-block-end: 0.25em;
+
+	cursor: pointer;
+	color: var(--light-gray);
+	text-transform: uppercase;
+	font-size: smaller;
+}
+
 summary {
 	cursor: pointer;
 }
@@ -67,4 +78,9 @@ a.clickable:hover {
 
 .clickable:disabled {
 	background-color: rgba(var(--gray-rgb), 0.3);
+}
+
+/* Bootstrap icons */
+svg.bi {
+	vertical-align: text-bottom;
 }
