--- src/routes/(hmf)/+layout.svelte
@@ -1,86 +1,0 @@
-<script>
-	import { page } from "$app/stores";
-
-	let path = $page.url.pathname;
-	$: path = $page.url.pathname;
-</script>
-
-<div class="hmf">
-	<header>
-		<nav>
-			<a href="/" class:active={path == "/"}>🏠 Home</a>
-			<a href="/about" class:active={path == "/about"}>👨🏻‍💻 About</a>
-			{#if path.startsWith("/projects")}
-				<a href="/projects" class="active">📂 Projects</a>
-			{:else}
-				<a href="/projects">📁 Projects</a>
-			{/if}
-			<a href="/blog" class:active={path.startsWith("/blog")}>📄 Blog</a>
-			<a href="/photos" class:active={path.startsWith("/photos")}>📷 Photos</a>
-		</nav>
-	</header>
-	<main>
-		<slot />
-	</main>
-	<footer>
-		<span>
-			Made with ❤️‍🔥 by <a href="/about">Felix</a> in
-			<a href="https://theländ.de" class="laend" rel="noopener noreferrer">
-				The Länd
-			</a>
-		</span>
-		<span>
-			<a href="/imprint">📜 Imprint</a>
-			|
-			<a href="https://github.com/felix-schindler/blog">Typo? Submit PR</a>
-		</span>
-	</footer>
-</div>
-
-<style lang="scss">
-	div.hmf {
-		header {
-			display: flex;
-			justify-content: space-around;
-		}
-
-		main {
-			padding: 0 1em;
-		}
-
-		footer {
-			display: flex;
-			align-items: center;
-			flex-direction: column-reverse;
-			gap: 0.75em;
-
-			@media (max-width: 1024px) {
-				justify-content: space-around;
-			}
-
-			a.laend {
-				text-decoration: none !important;
-
-				color: yellow;
-				background-color: black;
-
-				font-weight: 600;
-				text-transform: uppercase;
-				padding: 0.2em 0.4em;
-
-				transition: all 200ms;
-
-				&:hover {
-					color: black;
-					background-color: yellow;
-
-					box-shadow: 5px 5px;
-				}
-
-				&:active {
-					box-shadow: 0 0;
-				}
-			}
-		}
-	}
-</style>
