--- src/routes/(hmf)/photos/+page.svelte
@@ -1,9 +1,0 @@
-<h1>Photos</h1>
-<div class="grid">
-	<a href="/photos/germany" class="card clickable">
-		<h2>Germany</h2>
-	</a>
-	<a href="/photos/china" class="card clickable">
-		<h2>China</h2>
-	</a>
-</div>
