--- src/routes/(hmf)/projects/+page.svelte
@@ -1,13 +1,0 @@
-<h1>Projects</h1>
-<p>I always tend to delete my projects. These ones I won't delete that fast:</p>
-
-<div class="grid">
-	<div class="card">
-		<h2>Portfolio</h2>
-		<p>It's this website. It's gone through many iterations.</p>
-	</div>
-	<a href="/projects/tanuki" class="card clickable">
-		<h2>Tanuki for GitLab</h2>
-		<p>It's GitLab. On your phone.</p>
-	</a>
-</div>
