+++ src/routes/+layout.svelte
@@ -1,5 +1,83 @@
 <script>
 	import "$lib/style/main.css"; // Global styling
+	import { page } from "$app/stores";
+
+	let path;
+	$: path = $page.url.pathname;
+
+	let style;
+	$: path.includes("/photos/")
+		? (style = "padding: 0")
+		: (style = "padding: 0 1em");
 </script>
 
-<slot />
+<div class="hmf">
+	<header>
+		<nav>
+			<a href="/" class:active={path == "/"}>🏠 Home</a>
+			<a href="/about" class:active={path == "/about"}>👨🏻‍💻 About</a>
+			<a href="/content" class:active={path.startsWith("/content")}
+				>💾 Content</a
+			>
+		</nav>
+	</header>
+	<main {style}>
+		<slot />
+	</main>
+	<footer>
+		<span>
+			Made with ❤️‍🔥 by <a href="/about">Felix</a> in
+			<a href="https://theländ.de" class="laend" rel="noopener noreferrer">
+				The Länd
+			</a>
+		</span>
+		<span>
+			<a href="/imprint">📜 Imprint</a>
+			|
+			<a href="https://github.com/felix-schindler/blog">Typo? Submit PR</a>
+		</span>
+	</footer>
+</div>
+
+<style lang="scss">
+	div.hmf {
+		main {
+			padding: 0 1em;
+		}
+
+		footer {
+			display: flex;
+			align-items: center;
+			flex-direction: column-reverse;
+			gap: 0.75em;
+
+			@media (max-width: 1024px) {
+				justify-content: space-around;
+			}
+
+			a.laend {
+				text-decoration: none !important;
+
+				color: yellow;
+				background-color: black;
+
+				font-weight: 600;
+				text-transform: uppercase;
+				padding: 0.2em 0.4em;
+
+				transition: all 200ms;
+
+				&:hover {
+					color: black;
+					background-color: yellow;
+
+					box-shadow: 5px 5px;
+				}
+
+				&:active {
+					box-shadow: 0 0;
+				}
+			}
+		}
+	}
+</style>
