+++ src/routes/content/+layout.svelte
@@ -1,0 +1,15 @@
+<script>
+	import { page } from "$app/stores";
+	import Markdown from "$lib/components/Markdown.svelte";
+
+	let path;
+	$: path = $page.url.pathname;
+</script>
+
+{#if !path.includes("/photos/")}
+	<Markdown>
+		<slot />
+	</Markdown>
+{:else}
+	<slot />
+{/if}
