+++ src/routes/content/+page.svelte
@@ -1,0 +1,58 @@
+<h1>Content</h1>
+
+<h2 id="projects">Projects</h2>
+<p>I always tend to delete my projects. These ones I won't delete that fast:</p>
+<div class="grid">
+	<div class="card">
+		<h3>Portfolio</h3>
+		<p>It's this website. It's gone through many iterations.</p>
+	</div>
+	<a href="/content/projects/tanuki" class="card clickable">
+		<h3>Tanuki for GitLab</h3>
+		<p>It's GitLab. On your phone.</p>
+	</a>
+</div>
+
+<h2 id="photos">Photos</h2>
+<p>I like taking photos, however, I stopped carrying around a big camera.</p>
+<div class="grid">
+	<a href="/content/photos/germany" class="card clickable">
+		<h3>Germany</h3>
+	</a>
+	<a href="/content/photos/china" class="card clickable">
+		<h3>China</h3>
+	</a>
+</div>
+
+<h2 id="blog">Blog</h2>
+<p>thoughts and prayers</p>
+<div class="grid">
+	<a href="/content/blog/pgp" class="card clickable">
+		<h3>PGP / GPG and Git</h3>
+		<p>How to setup and use (on Mac)</p>
+	</a>
+	<a href="/content/blog/mc-perf" class="card clickable">
+		<h3>Minecraft Perf Mods</h3>
+		<p>Collection of some Minecraft Fabric performance mods</p>
+	</a>
+	<a href="/content/blog/reset" class="card clickable">
+		<h3>reset</h3>
+		<p>Hi, hello and welcome!</p>
+	</a>
+</div>
+
+<style lang="scss">
+	// Overwrite markdown style
+	a {
+		color: inherit !important;
+	}
+
+	a.clickable:hover {
+		text-decoration: none !important;
+	}
+
+	.card {
+		padding: 1em !important;
+		background-color: rgba(var(--gray-rgb), 0.2);
+	}
+</style>
